Added leader photos, fixed eos image, added eos banner to files
Added leadership image for Zoey and Vivian. Edited the name of the Eos photo as it was not case sensitive before (to possibly fix the issue of the image not appearing on the website). Added Eos banner image to the files, but I (elizabeth) am unsure of where the banner images are used in the code, so it was not implemented.